+++ finagle-base-http/src/main/scala/com/twitter/finagle/http/HttpClientTraceInitializer.scala
@@ -1,8 +1,24 @@
 package com.twitter.finagle.http
 
 import com.twitter.finagle
-import com.twitter.finagle.tracing.{Trace, TraceInitializerFilter}
+import com.twitter.finagle.tracing.{Trace, TraceInitializerFilter, Tracer}
 import com.twitter.finagle.{Filter, ServiceFactory, Stack}
+
+private[finagle] object HttpClientTraceInitializer {
+
+  def apply[Req, Rep](tracer: Tracer): Filter[Req, Rep, Req, Rep] =
+    Filter.mk[Req, Rep, Req, Rep] { (req, svc) =>
+      Trace.letTracerAndNextId(tracer) {
+        TraceInfo.setClientRequestHeaders(req.asInstanceOf[Request])
+        svc(req)
+      }
+    }
+
+  def typeAgnostic(tracer: Tracer): Filter.TypeAgnostic = new Filter.TypeAgnostic {
+    def toFilter[Req, Rep]: Filter[Req, Rep, Req, Rep] = apply(tracer)
+  }
+
+}
 
 private[finagle] class HttpClientTraceInitializer[Req <: Request, Rep]
     extends Stack.Module1[finagle.param.Tracer, ServiceFactory[Req, Rep]] {
@@ -13,13 +29,7 @@ private[finagle] class HttpClientTraceInitializer[Req <: Request, Rep]
     _tracer: finagle.param.Tracer,
     next: ServiceFactory[Req, Rep]
   ): ServiceFactory[Req, Rep] = {
-    val finagle.param.Tracer(tracer) = _tracer
-    val traceInitializer = Filter.mk[Req, Rep, Req, Rep] { (req, svc) =>
-      Trace.letTracerAndNextId(tracer) {
-        TraceInfo.setClientRequestHeaders(req)
-        svc(req)
-      }
-    }
+    val traceInitializer = HttpClientTraceInitializer[Req, Rep](_tracer.tracer)
     traceInitializer.andThen(next)
   }
 }